requirements.txt
fastapi
uvicorn
databases[postgresql]
sqlalchemy
pandas
python-multipart
python-jose
